Move history module from browser to common.
